Actions on the toolpanel: send, download submissions, reset are enabled for the code-quiz only
