+++ stepik-union/src/main/java/org/stepik/plugin/actions/step/InsertStepikDirectives.java
@@ -3,7 +3,6 @@ package org.stepik.plugin.actions.step;
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.KeyboardShortcut;
-import com.intellij.openapi.actionSystem.Presentation;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.fileEditor.FileEditorManager;
@@ -11,12 +10,12 @@ import com.intellij.openapi.keymap.KeymapUtil;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Pair;
 import com.intellij.openapi.vfs.VirtualFile;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.stepik.core.StepikProjectManager;
 import org.stepik.core.SupportedLanguages;
 import org.stepik.core.courseFormat.StepNode;
 import org.stepik.core.courseFormat.StudyNode;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.stepik.core.metrics.Metrics;
 import org.stepik.plugin.utils.DirectivesUtils;
 import org.stepik.plugin.utils.ReformatUtils;
@@ -31,7 +30,7 @@ import static org.stepik.plugin.utils.DirectivesUtils.removeAmbientCode;
 import static org.stepik.plugin.utils.DirectivesUtils.writeInToFile;
 
 
-public class InsertStepikDirectives extends AbstractStepAction {
+public class InsertStepikDirectives extends CodeQuizAction {
     private static final String SHORTCUT = "ctrl alt pressed R";
     private static final String ACTION_ID = "STEPIK.InsertStepikDirectives";
 
@@ -108,15 +107,5 @@ public class InsertStepikDirectives extends AbstractStepAction {
                 ReformatUtils.reformatSelectedEditor(project, document);
             }
         }
-    }
-
-    @Override
-    public void update(AnActionEvent e) {
-        super.update(e);
-        Presentation presentation = e.getPresentation();
-        StudyNode<?, ?> selectedNode = StepikProjectManager.getSelected(e.getProject());
-        boolean enabled = presentation.isEnabled();
-        boolean canEnabled = (selectedNode instanceof StepNode) && (((StepNode) selectedNode).getType() == CODE);
-        presentation.setEnabled(enabled && canEnabled);
     }
 }
